Create helper module
This way the helper functions can not only be used in osbrain tests but
also in other modules tests (i.e.: osmarkets).
