+++ osbrain/tests/test_agent_async_requests.py
@@ -5,10 +5,10 @@ import time
 
 from osbrain import run_agent
 from osbrain import run_logger
+from osbrain.helper import sync_agent_logger
+from osbrain.helper import logger_received
 
 from common import nsproxy  # pragma: no flakes
-from common import sync_agent_logger
-from common import logger_received
 
 
 def on_error(agent):
