+++ osbrain/tests/test_agent_sync_publications.py
@@ -8,11 +8,11 @@ import pytest
 from osbrain import Agent
 from osbrain import run_agent
 from osbrain import run_logger
+from osbrain.helper import sync_agent_logger
+from osbrain.helper import logger_received
+from osbrain.helper import wait_agent_attr
 
 from common import nsproxy  # pragma: no flakes
-from common import sync_agent_logger
-from common import logger_received
-from common import wait_agent_attr
 
 
 class Server(Agent):
